requirements.txt
# Core framework
streamlit==1.41.1

# Data handling
pandas==2.2.2
altair==5.3.0
pillow==10.4.0
openpyxl==3.1.5
numpy==1.26.4

# Firebase Admin SDK
firebase-admin==6.5.0

# Protobuf (compatible version)
protobuf==4.25.3

# QR code + ML utils
qrcode==7.4.2
joblib==1.3.2
